+++ framework/domain/src/main/java/is/codion/framework/domain/entity/IdentityKeyGenerator.java
@@ -28,7 +28,7 @@ final class IdentityKeyGenerator implements KeyGenerator {
       if (!generatedKeys.next()) {
         throw new SQLException("Identity key generator returned no generated keys", DatabaseConnection.SQL_STATE_NO_DATA);
       }
-      ColumnDefinition<Object> column = (ColumnDefinition<Object>) entity.definition().primaryKey().columnDefinitions().get(0);
+      ColumnDefinition<Object> column = (ColumnDefinition<Object>) entity.definition().primaryKey().definitions().get(0);
       // must fetch value by column name, since some databases (PostgreSQL for example), return all columns, not just generated ones
       entity.put(column.attribute(), column.prepareValue(generatedKeys.getObject(column.name())));
     }